judgecore/requirements.txt
aio_pika
python-dotenv~=0.18.0
pydantic